pyproject.toml
[project]
name = "agentes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.7.2",
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "google-genai>=1.25.0",
    "groq>=0.30.0",
    "openai>=1.96.0",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.41",
    "tavily-python>=0.7.9",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
]
